requirements.txt
spotipy==2.25.1
python-dotenv==1.0.0
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.3